% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{download_doc}
\alias{download_doc}
\title{Download document}
\usage{
download_doc(
  id,
  destdir = getwd(),
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{id}{integer}

\item{destdir}{string}

\item{base_url}{string}

\item{api_key}{string}
}
\value{
no output, called for side effects.
}
\description{
Downloads a document.
}
\examples{
\dontrun{
download_doc(17)
}
}
