% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{get_all_docs_df}
\alias{get_all_docs_df}
\title{Get all documents dataframe}
\usage{
get_all_docs_df(base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{base_url}{string}

\item{api_key}{string}
}
\value{
data frame
}
\description{
Get a dataframe with information about all the documents on a
Paperless-ngx instance (one row per document).
}
\examples{
\dontrun{
get_all_docs_df()
}
}
