% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API key from environment variable}
\usage{
get_api_key()
}
\value{
string (if variable set)
}
\description{
Gets the API key to a Paperless-ngx instance from an
environment variable. Requires the environment variable
\code{PAPERLESS_API_KEY} to be set in \code{.Renviron}.
}
\examples{
\dontrun{
get_api_key()
}
}
