% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_base_url}
\alias{get_base_url}
\title{Get base URL from environment variable}
\usage{
get_base_url()
}
\value{
string (if variable set)
}
\description{
Gets the base URL to a Paperless-ngx instance from an
environment variable. Requires the environment variable
\code{PAPERLESS_BASE_URL} to be set in \code{.Renviron}.
}
\examples{
\dontrun{
get_base_url()
}
}
