% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_corrs}
\alias{get_corrs}
\title{Get correspondents info}
\usage{
get_corrs(base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{base_url}{string}

\item{api_key}{string}
}
\value{
list
}
\description{
Get information on all correspondents in the Paperless-ngx
instance in list form.
}
\examples{
\dontrun{
get_corrs()
}
}
