% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{get_doc_ids}
\alias{get_doc_ids}
\title{Get document ids}
\usage{
get_doc_ids(base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{base_url}{string}

\item{api_key}{string}
}
\value{
integer vector
}
\description{
Gets the document ids for all documents on a Paperless-ngx
instance.
}
\examples{
\dontrun{
get_doc_ids()
}
}
