% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{get_doc_info}
\alias{get_doc_info}
\title{Get document info}
\usage{
get_doc_info(id, base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{id}{integer}

\item{base_url}{string}

\item{api_key}{string}
}
\value{
list
}
\description{
Get information about a document in list form.
}
\examples{
\dontrun{
get_doc_info()
}
}
