% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{get_doc_info_df}
\alias{get_doc_info_df}
\title{Get document info dataframe}
\usage{
get_doc_info_df(
  id,
  base_url = get_base_url(),
  api_key = get_api_key(),
  n_tags = 5
)
}
\arguments{
\item{id}{integer}

\item{base_url}{string}

\item{api_key}{string}

\item{n_tags}{integer. The number of tag columns you want in the data frame.}
}
\value{
data frame
}
\description{
Get information about a document in dataframe form.
}
\examples{
\dontrun{
get_doc_info_df(17)
}
}
