% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{req_auth}
\alias{req_auth}
\title{Authenticate request}
\usage{
req_auth(base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{base_url}{string}

\item{api_key}{string}
}
\value{
httr2_request
}
\description{
Creates an httr2 authentication request.
}
\examples{
\dontrun{
req_auth()
}
}
