% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{search_docs}
\alias{search_docs}
\title{Search documents}
\usage{
search_docs(query, base_url = get_base_url(), api_key = get_api_key())
}
\arguments{
\item{query}{string}

\item{base_url}{string}

\item{api_key}{string}
}
\value{
integer vector
}
\description{
Search a Paperless-ngx instance. Returns a vector of document
ids if the term is found.
}
\examples{
\dontrun{
search_docs("london")
}
}
