% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{upload_doc}
\alias{upload_doc}
\title{Upload document}
\usage{
upload_doc(
  path,
  title = NULL,
  created = NULL,
  correspondent = NULL,
  document_type = NULL,
  storage_path = NULL,
  tags = NULL,
  archive_serial_number = NULL,
  owner = NULL,
  set_permissions = NULL,
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{path}{string}

\item{title}{string}

\item{created}{string}

\item{correspondent}{integer}

\item{document_type}{integer}

\item{storage_path}{string}

\item{tags}{list}

\item{archive_serial_number}{integer}

\item{owner}{integer}

\item{set_permissions}{string}

\item{base_url}{string}

\item{api_key}{string}
}
\value{
no output, called for side effects.
}
\description{
Add a document to a Paperless-ngx instance.
}
\examples{
\dontrun{
upload_doc("scan.pdf")
}
}
